Generate a bundle file for all Microsoft CA's, as per CCADB
This does revoke trust for two certificates now present in authenticode-legacy.pem (this is not trusted by default), as these are not in the CCADB.

Note that we do NOT currently implement policies as defined in the authroot.stl file that Microsoft uses. This means that we will now trust certificates that have been disabled by Microsoft for some reason, but still need to be trusted for previously issued certificates. (If we were acting as a simple 'now' program, we would ignore those.) Those certificates are shown as 'revoked' in the Certificate Manager of Windows.